Migrate AssetGalleryFieldTest to AssetAdminTest
Removed a few irrelevant tests about field instanciations
