+++ tests/php/AssetAdminTest.yml
@@ -1,6 +1,15 @@
+Member:
+  member1:
+    FirstName: Christopher
+    LastName: Pitt
+  member2:
+    FirstName: Damian
+    LastName: Mooyman
 Folder:
   folder1:
-    Name: AssetAdminTest
+    Title: Folder1
+  folder2:
+    Title: Folder2
 File:
   file1:
     Title: File1
@@ -10,3 +19,34 @@ File:
     Title: File2
     Name: file2.txt
     ParentID: =>Folder.folder1
+SilverStripe\AssetAdmin\Tests\AssetAdminTest_File:
+  file1:
+    Created: 2015-08-19 00:00:00
+    LastUpdated: 2015-08-27 00:00:00
+    Owner: =>Member.member1
+    Parent: =>Folder.folder1
+    Name: file1.jpg
+    Title: The First File
+  file2:
+    Created: 2015-08-20 00:00:00
+    LastUpdated: 2015-08-28 00:00:00
+    Owner: =>Member.member2
+    Parent: =>Folder.folder2
+    Name: file2.jpg
+    Title: The Second File
+  file3:
+    Created: 2015-08-21 00:00:00
+    LastUpdated: 2015-08-29 00:00:00
+    Owner: =>Member.member2
+    Parent: =>Folder.folder2
+    Name: file3.jpg
+    Title: The Third File
+  disallowCanView:
+    Parent: =>Folder.folder1
+    Name: disallowCanView.txt
+  disallowCanEdit:
+    Parent: =>Folder.folder1
+    Name: disallowCanEdit.txt
+  disallowCanDelete:
+    Parent: =>Folder.folder1
+    Name: disallowCanDelete.txt
